zone company check and exptioin

+++ src/zone/zone.controller.ts
@@ -1,4 +1,4 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete, UseGuards,Request } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, UseGuards,Request, BadRequestException, HttpException, HttpStatus } from '@nestjs/common';
 import { ZoneService } from './zone.service';
 import { CreateZoneDto } from './dto/create-zone.dto';
 import { UpdateZoneDto } from './dto/update-zone.dto';
@@ -7,30 +7,48 @@ import { AuthGuard } from 'src/auth/auth.guard';
 @Controller('zone')
 export class ZoneController {
   constructor(private readonly zoneService: ZoneService) {}
-
+  @UseGuards(AuthGuard)
   @Post()
-  create(@Body() createZoneDto: CreateZoneDto) {
-    return this.zoneService.create(createZoneDto);
+  async create(@Request() req ,@Body() createZoneDto: CreateZoneDto) {
+    if (req.user.selectCompany !== createZoneDto.companyId)
+    {
+        console.error('invalid company');
+        throw new HttpException({
+              status: HttpStatus.BAD_REQUEST,
+              error: 'BAD_REQUEST',
+            }, HttpStatus.BAD_REQUEST, {
+              cause: 'invalid company one'
+            });
+    }
+    return await this.zoneService.create(createZoneDto);
   }
   @UseGuards(AuthGuard)
   @Get()
-  findAll(@Request() req  ) {
-    console.log(req.user)
-    return this.zoneService.findAll(req.user.selectCompany);
+  async findAll(@Request() req  ) {    
+    return await this.zoneService.findAll(req.user.selectCompany);
   }
 
   @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.zoneService.findOne(+id);
+  async findOne(@Request() req ,@Param('id') id: string) {
+    
+    const zone = await this.zoneService.findOne(+id);
+    if (req.user.selectCompany !== zone.company.id)
+        throw BadRequestException;
+    return zone;
   }
 
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateZoneDto: UpdateZoneDto) {
+  async update(@Request() req ,@Param('id') id: string, @Body() updateZoneDto: UpdateZoneDto) {
+    if (req.user.selectCompany !== updateZoneDto.companyId)
+        throw BadRequestException;
     return this.zoneService.update(+id, updateZoneDto);
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  async remove(@Request() req, @Param('id') id: string) {
+    const zone = await  this.zoneService.findOne(+id);
+    if (req.user.selectCompany !== zone.company.id)
+        throw BadRequestException;
     return this.zoneService.remove(+id);
   }
 }
